+++ zoet-macro/src/traits.rs
@@ -48,7 +48,8 @@ pub(crate) fn get_trait_fn(ident: &Ident) -> Result<GenFn> {
         // "GlobalAlloc" => Requires two functions to be implemented.
 
         // std::any
-        // "Any" => doable.
+        // "Any" => possible, but it's already automatically-implemented by the compiler.
+        // "Provider" => nightly.
 
         // std::borrow
         "Borrow" => |f| borrow_shaped(f, &pq!(::core::borrow::Borrow), &pq!(borrow)),
@@ -77,7 +78,8 @@ pub(crate) fn get_trait_fn(ident: &Ident) -> Result<GenFn> {
         "Default" => default,
 
         // std::error
-        // "Error" => Generate `source()`?
+        // "Error" => possible because all methods are defaulted, but which method should we
+        // implement? Use e.g. thiserror instead.
 
         // std::fmt
         "Binary" => |f| debug_shaped(f, &pq!(::core::fmt::Binary)),
@@ -93,10 +95,10 @@ pub(crate) fn get_trait_fn(ident: &Ident) -> Result<GenFn> {
 
         // std::future
         "Future" => future,
-        // "IntoFuture" ?
+        "IntoFuture" => into_future,
 
         // std::hash
-        // "BuildHasher"?
+        // "BuildHasher" => possible, arguably TODO.
         "Hash" => hash,
         // "Hasher" => Requires two functions to be implemented.
 
@@ -183,21 +185,15 @@ fn borrow_shaped(func: GenIn, trait_name: &Path, method_name: &Ident) -> Result<
     let FunctionArgs {
         input,
         output,
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.unwrap_ref_param(0)?.unary()?.unwrap_ref_return()?;
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
         derive_span =>
             #extra_attrs
         impl #generics #trait_name <#output> for #input #where_clause {
-             fn #method_name (&self) -> &#output {
+            #[inline]
+            fn #method_name (&self) -> &#output {
                 #to_call(self)
             }
         }
@@ -208,21 +204,15 @@ fn borrow_mut_shaped(func: GenIn, trait_name: &Path, method_name: &Ident) -> Res
     let FunctionArgs {
         input,
         output,
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.unwrap_mut_param(0)?.unary()?.unwrap_mut_return()?;
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
         derive_span =>
             #extra_attrs
         impl #generics #trait_name <#output> for #input #where_clause {
-             fn #method_name (&mut self) -> &mut #output {
+            #[inline]
+            fn #method_name (&mut self) -> &mut #output {
                 #to_call(self)
             }
         }
@@ -233,14 +223,7 @@ fn to_owned(func: GenIn) -> Result<TokenStream> {
     let FunctionArgs {
         input,
         output,
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.unwrap_ref_param(0)?.unary()?.has_return()?;
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
@@ -248,7 +231,8 @@ fn to_owned(func: GenIn) -> Result<TokenStream> {
             #extra_attrs
         impl #generics ::zoet::__alloc::borrow::ToOwned for #input #where_clause {
             type Owned = #output;
-             fn to_owned(&self) -> Self::Owned {
+            #[inline]
+            fn to_owned(&self) -> Self::Owned {
                 #to_call(self)
             }
         }
@@ -259,21 +243,15 @@ fn clone(func: GenIn) -> Result<TokenStream> {
     let FunctionArgs {
         input,
         output,
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.unwrap_ref_param(0)?.unary()?.has_return()?; // TODO: create and use a .self_return() instead?
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
         derive_span =>
             #extra_attrs
         impl #generics ::core::clone::Clone for #input #where_clause {
-             fn clone(&self) -> #output {
+            #[inline]
+            fn clone(&self) -> #output {
                 #to_call(self)
             }
         }
@@ -328,14 +306,7 @@ fn ord(func: GenIn) -> Result<TokenStream> {
         FunctionArgs {
             input: (lhs, rhs),
             output,
-            meta:
-                FunctionMeta {
-                    derive_span,
-                    generics,
-                    to_call,
-                    extra_attrs,
-                    ..
-                },
+            meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
         },
     ) = compare_shape(func)?;
 
@@ -354,7 +325,7 @@ fn ord(func: GenIn) -> Result<TokenStream> {
         derive_span =>
             #extra_attrs
         impl #generics ::core::cmp::Ord for #lhs #where_clause {
-             fn cmp(&self, other: &#rhs) -> ::core::cmp::Ordering {
+            fn cmp(&self, other: &#rhs) -> ::core::cmp::Ordering {
                 #fn_body
             }
         }
@@ -366,14 +337,7 @@ fn partial_eq(func: GenIn) -> Result<TokenStream> {
         shape,
         FunctionArgs {
             input: (lhs, rhs),
-            meta:
-                FunctionMeta {
-                    derive_span,
-                    generics,
-                    to_call,
-                    extra_attrs,
-                    ..
-                },
+            meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
             ..
         },
     ) = compare_shape(func)?;
@@ -396,7 +360,8 @@ fn partial_eq(func: GenIn) -> Result<TokenStream> {
         derive_span =>
             #extra_attrs
         impl #generics ::core::cmp::PartialEq<#rhs> for #lhs #where_clause {
-             fn eq(&self, other: &#rhs) -> ::core::primitive::bool {
+            #[inline]
+            fn eq(&self, other: &#rhs) -> ::core::primitive::bool {
                 #fn_body
             }
         }
@@ -409,14 +374,7 @@ fn partial_ord(func: GenIn) -> Result<TokenStream> {
         FunctionArgs {
             input: (lhs, rhs),
             output,
-            meta:
-                FunctionMeta {
-                    derive_span,
-                    generics,
-                    to_call,
-                    extra_attrs,
-                    ..
-                },
+            meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
         },
     ) = compare_shape(func)?;
 
@@ -438,7 +396,8 @@ fn partial_ord(func: GenIn) -> Result<TokenStream> {
         derive_span =>
             #extra_attrs
         impl #generics ::core::cmp::PartialOrd<#rhs> for #lhs #where_clause {
-             fn partial_cmp(&self, other: &#rhs) -> ::core::option::Option<::core::cmp::Ordering> {
+            #[inline]
+            fn partial_cmp(&self, other: &#rhs) -> ::core::option::Option<::core::cmp::Ordering> {
                 #fn_body
             }
         }
@@ -450,21 +409,15 @@ fn from(func: GenIn) -> Result<TokenStream> {
     let FunctionArgs {
         input,
         output,
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.unary()?.has_return()?; // TODO: .self_return()?
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
         derive_span =>
             #extra_attrs
         impl #generics ::core::convert::From<#input> for #output #where_clause {
-             fn from(value: #input) -> Self {
+            #[inline]
+            fn from(value: #input) -> Self {
                 #to_call(value)
             }
         }
@@ -475,21 +428,15 @@ fn into(func: GenIn) -> Result<TokenStream> {
     let FunctionArgs {
         input,
         output,
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.unary()?.has_return()?;
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
         derive_span =>
             #extra_attrs
         impl #generics ::core::convert::Into<#output> for #input #where_clause {
-             fn into(self) -> #output {
+            #[inline]
+            fn into(self) -> #output {
                 #to_call(self)
             }
         }
@@ -501,14 +448,7 @@ fn try_from(func: GenIn) -> Result<TokenStream> {
     let FunctionArgs {
         input,
         output: (output, err),
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.unary()?.unwrap_result_return()?;
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
@@ -516,7 +456,8 @@ fn try_from(func: GenIn) -> Result<TokenStream> {
             #extra_attrs
         impl #generics ::core::convert::TryFrom<#input> for #output #where_clause {
             type Error = #err;
-             fn try_from(value: #input) -> ::core::result::Result<Self, Self::Error> {
+            #[inline]
+            fn try_from(value: #input) -> ::core::result::Result<Self, Self::Error> {
                 #to_call(value)
             }
         }
@@ -527,14 +468,7 @@ fn try_into(func: GenIn) -> Result<TokenStream> {
     let FunctionArgs {
         input,
         output: (output, err),
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.unary()?.unwrap_result_return()?;
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
@@ -542,7 +476,8 @@ fn try_into(func: GenIn) -> Result<TokenStream> {
             #extra_attrs
         impl #generics ::core::convert::TryInto<#output> for #input #where_clause {
             type Error = #err;
-             fn try_into(self) -> ::core::result::Result<#output, Self::Error> {
+            #[inline]
+            fn try_into(self) -> ::core::result::Result<#output, Self::Error> {
                 #to_call(self)
             }
         }
@@ -553,21 +488,15 @@ fn default(func: GenIn) -> Result<TokenStream> {
     let FunctionArgs {
         input: (),
         output,
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.nullary()?.has_return()?;
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
         derive_span =>
             #extra_attrs
         impl #generics ::core::default::Default for #output #where_clause {
-             fn default() -> Self {
+            #[inline]
+            fn default() -> Self {
                 #to_call()
             }
         }
@@ -578,14 +507,7 @@ fn debug_shaped(func: GenIn, trait_name: &Path) -> Result<TokenStream> {
     let FunctionArgs {
         input: (obj, formatter),
         output,
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func
         .unwrap_ref_param(0)?
         .unwrap_mut_param(1)?
@@ -596,7 +518,8 @@ fn debug_shaped(func: GenIn, trait_name: &Path) -> Result<TokenStream> {
         derive_span =>
             #extra_attrs
         impl #generics #trait_name for #obj #where_clause {
-             fn fmt(&self, f: &mut #formatter) -> #output {
+            #[inline]
+            fn fmt(&self, f: &mut #formatter) -> #output {
                 #to_call(self, f)
             }
         }
@@ -607,14 +530,7 @@ fn write(func: GenIn) -> Result<TokenStream> {
     let FunctionArgs {
         input: (obj, str),
         output,
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func
         .unwrap_mut_param(0)?
         .unwrap_ref_param(1)?
@@ -625,7 +541,8 @@ fn write(func: GenIn) -> Result<TokenStream> {
         derive_span =>
             #extra_attrs
         impl #generics ::core::fmt::Write for #obj #where_clause {
-             fn write_str(&mut self, s: &#str) -> #output {
+            #[inline]
+            fn write_str(&mut self, s: &#str) -> #output {
                 #to_call(self, s)
             }
         }
@@ -637,14 +554,7 @@ fn future(func: GenIn) -> Result<TokenStream> {
     let FunctionArgs {
         input: (obj, _),
         output,
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func
         .unwrap_param(0, "Pin")?
         .unwrap_mut_param(0)?
@@ -657,6 +567,7 @@ fn future(func: GenIn) -> Result<TokenStream> {
         impl #generics ::core::future::Future for #obj #where_clause {
             type Output = #output;
 
+            #[inline]
             fn poll(
                 self: ::core::pin::Pin<&mut Self>,
                 cx: &mut ::core::task::Context
@@ -667,19 +578,35 @@ fn future(func: GenIn) -> Result<TokenStream> {
     })
 }
 
+/// `fn(self) -> impl Future<Output=T>`
+fn into_future(func: GenIn) -> Result<TokenStream> {
+    let FunctionArgs {
+        input,
+        output,
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
+    } = func.unary()?.has_return()?;
+    let where_clause = &generics.where_clause;
+    Ok(quote_spanned! {
+        derive_span =>
+            #extra_attrs
+        impl #generics ::core::future::IntoFuture for #input #where_clause {
+            type IntoFuture = #output;
+            type Output = <Self::IntoFuture as ::core::future::Future>::Output;
+
+            #[inline]
+             fn into_future(self) -> Self::IntoFuture {
+                 #to_call(self)
+             }
+        }
+    })
+}
+
 /// (Signature is particularly trait-specific.)
 fn hash(func: GenIn) -> Result<TokenStream> {
     let FunctionArgs {
         input: (obj, _hasher),
         output: (),
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func
         .unwrap_ref_param(0)?
         .unwrap_mut_param(1)?
@@ -690,7 +617,8 @@ fn hash(func: GenIn) -> Result<TokenStream> {
         derive_span =>
             #extra_attrs
         impl #generics ::core::hash::Hash for #obj #where_clause {
-             fn hash<H: ::core::hash::Hasher>(&self, state: &mut H) {
+            #[inline]
+            fn hash<H: ::core::hash::Hasher>(&self, state: &mut H) {
                 #to_call(self, state);
             }
         }
@@ -703,14 +631,7 @@ fn into_iterator(func: GenIn) -> Result<TokenStream> {
     let FunctionArgs {
         input,
         output,
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.unary()?.has_return()?;
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
@@ -719,7 +640,8 @@ fn into_iterator(func: GenIn) -> Result<TokenStream> {
         impl #generics ::core::iter::IntoIterator for #input #where_clause {
             type IntoIter = #output;
             type Item = <Self::IntoIter as ::core::iter::Iterator>::Item;
-             fn into_iter(self) -> Self::IntoIter {
+            #[inline]
+            fn into_iter(self) -> Self::IntoIter {
                 #to_call(self)
             }
         }
@@ -731,14 +653,7 @@ fn iterator(func: GenIn) -> Result<TokenStream> {
     let FunctionArgs {
         input,
         output,
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.unwrap_mut_param(0)?.unary()?.unwrap_return("Option")?;
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
@@ -746,7 +661,8 @@ fn iterator(func: GenIn) -> Result<TokenStream> {
             #extra_attrs
         impl #generics ::core::iter::Iterator for #input #where_clause {
             type Item = #output;
-             fn next(&mut self) -> Option<Self::Item> {
+            #[inline]
+            fn next(&mut self) -> Option<Self::Item> {
                 #to_call(self)
             }
         }
@@ -764,14 +680,7 @@ fn add_shaped(func: GenIn, trait_name: &Path, method_name: &Ident) -> Result<Tok
     let FunctionArgs {
         input: (lhs, rhs),
         output,
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.binary()?.has_return()?;
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
@@ -779,7 +688,8 @@ fn add_shaped(func: GenIn, trait_name: &Path, method_name: &Ident) -> Result<Tok
             #extra_attrs
         impl #generics #trait_name<#rhs> for #lhs #where_clause {
             type Output = #output;
-             fn #method_name(self, rhs: #rhs) -> Self::Output {
+            #[inline]
+            fn #method_name(self, rhs: #rhs) -> Self::Output {
                 #to_call(self, rhs)
             }
         }
@@ -792,14 +702,7 @@ fn add_from_add_assign_shaped(
     let FunctionArgs {
         input: (lhs, rhs),
         output: (),
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.unwrap_mut_param(0)?.binary()?.default_return()?;
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
@@ -807,7 +710,8 @@ fn add_from_add_assign_shaped(
             #extra_attrs
         impl #generics #trait_name<#rhs> for #lhs #where_clause {
             type Output = #lhs;
-             fn #method_name(mut self, rhs: #rhs) -> #lhs {
+            #[inline]
+            fn #method_name(mut self, rhs: #rhs) -> #lhs {
                 #to_call(&mut self, rhs) ; self
             }
         }
@@ -819,21 +723,15 @@ fn add_assign_shaped(func: GenIn, trait_name: &Path, method_name: &Ident) -> Res
     let FunctionArgs {
         input: (lhs, rhs),
         output: (),
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.unwrap_mut_param(0)?.binary()?.default_return()?;
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
         derive_span =>
             #extra_attrs
         impl #generics #trait_name<#rhs> for #lhs #where_clause {
-             fn #method_name(&mut self, rhs: #rhs) {
+            #[inline]
+            fn #method_name(&mut self, rhs: #rhs) {
                 #to_call(self, rhs);
             }
         }
@@ -844,21 +742,15 @@ fn drop(func: GenIn) -> Result<TokenStream> {
     let FunctionArgs {
         input,
         output: (),
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.unwrap_mut_param(0)?.unary()?.default_return()?;
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
         derive_span =>
             #extra_attrs
         impl #generics ::core::ops::Drop for #input #where_clause {
-             fn drop(&mut self) {
+            #[inline]
+            fn drop(&mut self) {
                 #to_call(self);
             }
         }
@@ -900,14 +792,7 @@ fn index(func: GenIn) -> Result<TokenStream> {
     let FunctionArgs {
         input: (coll, idx),
         output,
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.unwrap_ref_param(0)?.binary()?.unwrap_ref_return()?;
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
@@ -915,6 +800,7 @@ fn index(func: GenIn) -> Result<TokenStream> {
             #extra_attrs
         impl #generics ::core::ops::Index<#idx> for #coll #where_clause {
             type Output = #output;
+            #[inline]
              fn index(&self, index: #idx) -> &Self::Output {
                 #to_call(self, index)
             }
@@ -925,14 +811,7 @@ fn index(func: GenIn) -> Result<TokenStream> {
 fn index_mut(func: GenIn) -> Result<TokenStream> {
     let FunctionArgs {
         input: (coll, idx),
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
         ..
     } = func.unwrap_mut_param(0)?.binary()?.unwrap_mut_return()?;
     let where_clause = &generics.where_clause;
@@ -940,6 +819,7 @@ fn index_mut(func: GenIn) -> Result<TokenStream> {
         derive_span =>
             #extra_attrs
         impl #generics ::core::ops::IndexMut<#idx> for #coll #where_clause {
+            #[inline]
              fn index_mut(&mut self, index: #idx) -> &mut Self::Output {
                 #to_call(self, index)
             }
@@ -951,14 +831,7 @@ fn neg_shaped(func: GenIn, trait_name: &Path, method_name: &Ident) -> Result<Tok
     let FunctionArgs {
         input,
         output,
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.unary()?.has_return()?;
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
@@ -966,6 +839,7 @@ fn neg_shaped(func: GenIn, trait_name: &Path, method_name: &Ident) -> Result<Tok
             #extra_attrs
         impl #generics #trait_name for #input #where_clause {
             type Output = #output;
+            #[inline]
              fn #method_name(self) -> Self::Output {
                 #to_call(self)
             }
@@ -977,14 +851,7 @@ fn deref(func: GenIn) -> Result<TokenStream> {
     let FunctionArgs {
         input,
         output,
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.unwrap_ref_param(0)?.unary()?.unwrap_ref_return()?;
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
@@ -992,6 +859,7 @@ fn deref(func: GenIn) -> Result<TokenStream> {
             #extra_attrs
         impl #generics ::core::ops::Deref for #input #where_clause {
             type Target = #output;
+            #[inline]
              fn deref(&self) -> &Self::Target {
                 #to_call(self)
             }
@@ -1003,20 +871,14 @@ fn deref_mut(func: GenIn) -> Result<TokenStream> {
     let FunctionArgs {
         input,
         output,
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.unwrap_mut_param(0)?.unary()?.unwrap_mut_return()?;
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
         derive_span =>
             #extra_attrs
         impl #generics ::core::ops::DerefMut for #input #where_clause {
+            #[inline]
              fn deref_mut(&mut self) -> &mut #output {
                 #to_call(self)
             }
@@ -1029,14 +891,7 @@ fn from_str(func: GenIn) -> Result<TokenStream> {
     let FunctionArgs {
         input,
         output: (output, err),
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.check_ref_param(0)?.unary()?.unwrap_result_return()?;
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
@@ -1044,6 +899,7 @@ fn from_str(func: GenIn) -> Result<TokenStream> {
             #extra_attrs
         impl #generics ::core::str::FromStr for #output #where_clause {
             type Err = #err;
+            #[inline]
              fn from_str(s: #input) -> ::core::result::Result<Self, Self::Err> {
                 #to_call(s)
             }
@@ -1056,20 +912,14 @@ fn to_string(func: GenIn) -> Result<TokenStream> {
     let FunctionArgs {
         input,
         output,
-        meta:
-            FunctionMeta {
-                derive_span,
-                generics,
-                to_call,
-                extra_attrs,
-                ..
-            },
+        meta: FunctionMeta { derive_span, generics, to_call, extra_attrs, .. },
     } = func.unwrap_ref_param(0)?.unary()?.has_return()?;
     let where_clause = &generics.where_clause;
     Ok(quote_spanned! {
         derive_span =>
             #extra_attrs
         impl #generics ::zoet::__alloc::string::ToString for #input #where_clause {
+            #[inline]
              fn to_string(&self) -> #output {
                 #to_call(self)
             }